# docs/swagger.yaml
basePath: /kitchencontrol/api/v1
definitions:
...
  usecase.CheckPaymentOutputDTO:
    properties:
      gatewayTransactionID:
        type: string
      orderID:
        type: string
      paymentApproved:
        type: boolean
      status:
        type: string
    type: object
  usecase.CheckoutConfirmationInputDTO:
    properties:
      order_id:
        type: string
      status:
        type: string
    required:
    - order_id
    - status
    type: object
  usecase.CheckoutConfirmationOutputDTO:
    properties:
      checkout_id:
        type: string
      gateway_transaction_id:
        type: string
      order_id:
        type: string
      qrcode:
        type: string
      status:
        type: string
    required:
    - checkout_id
    - gateway_transaction_id
    - order_id
    - qrcode
    - status
    type: object
...
  usecase.CheckoutReprocessingInputDTO:
    properties:
      notification_url:
        description: webhook para receber a confirmação do pagamento
        type: string
      sponsor_id:
        type: string
    type: object
  usecase.CheckoutReprocessingOutputDTO:
    properties:
      checkout_id:
        type: string
      gateway_transaction_id:
        type: string
      order_id:
        type: string
      qrcode:
        type: string
      status:
        type: string
    required:
    - checkout_id
    - gateway_transaction_id
    - order_id
    - qrcode
    - status
    type: object
  usecase.CustomerFindAllOutputDTO:
    properties:
      cpf:
        type: string
      email:
        type: string
      name:
        type: string
    type: object
  usecase.CustomerFindByCpfOutputDTO:
    properties:
      cpf:
        type: string
      email:
        type: string
      name:
        type: string
    type: object
  usecase.CustomerRegisterInputDTO:
    properties:
      cpf:
        type: string
      email:
        type: string
      name:
        type: string
    type: object
  usecase.CustomerRegisterOutputDTO:
    properties:
      cpf:
        type: string
      email:
        type: string
      name:
        type: string
    type: object
  usecase.CustomerUpdateInputDTO:
    properties:
      cpf:
        type: string
      email:
        type: string
      name:
        type: string
    type: object
  usecase.CustomerUpdateOutputDTO:
    properties:
      cpf:
        type: string
      email:
        type: string
      name:
        type: string
    type: object
...
  usecase.KitchenCookingInputDTO:
    properties:
      order_id:
        type: string
    type: object
  usecase.KitchenCookingOutputDTO:
    properties:
      customer_id:
        type: string
      estimated_time:
        type: string
      items:
        items:
          type: string
        type: array
      kitchen_id:
        type: string
      order_id:
        type: string
      queue:
        type: string
      status:
        type: string
    type: object
...
  usecase.KitchenFindAllAOutputDTO:
    properties:
      created_at:
        type: string
      estimated_time:
        type: string
      id:
        type: string
      order_id:
        type: string
      queue:
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is fiap fase 2 challenge project.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Fiap Fase 2 Challenge Clean Arch API - 9SOAT
  version: "1.0"
paths:
...
  /checkouts/{id}/check/payment:
    get:
      consumes:
      - application/json
      description: 'Get details of an Checkout and Status of Payment by Order id.
        Req #2 - Consultar status de pagamento pedido, que informa se o pagamento
        foi aprovado ou não.'
      parameters:
      - description: Order id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usecase.CheckPaymentOutputDTO'
        "400":
          description: Order ID is required
          schema:
            type: string
        "404":
          description: Order | Checkout not found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Check Payment by Order ID
      tags:
      - Checkouts
  /checkouts/confirmation/payment:
    post:
      consumes:
      - application/json
      description: 'Confirma o pagamento do cliente, via fake checkout nesse momento,
        e libera o pedido para preparação. A ordem muda de status nesse momento, para
        pagamento aprovado. Req #1 - Webhook para receber confirmação de pagamento
        aprovado ou recusado. A implementação deve ser clara quanto ao Webhook.'
      parameters:
      - description: Webhook para finalizar o Checkout
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/usecase.CheckoutConfirmationInputDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usecase.CheckoutConfirmationOutputDTO'
        "400":
          description: invalid data
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Webhook to confirm payment
      tags:
      - Checkouts
  /checkouts/reprocessing/payment:
    post:
      consumes:
      - application/json
      description: Reprocesso o pagamento. Ordens em checkout aprovado, com checkout
        criado podem ser reprocessadas. O reprocessamento é feito no gateway de pagamento.
        O checkout é atualizado com o status do pagamento e o pedido é notificado.
      parameters:
      - description: reprocessa o Checkout
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/usecase.CheckoutReprocessingInputDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usecase.CheckoutReprocessingOutputDTO'
        "400":
          description: invalid data
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Reprocessa o Checkout
      tags:
      - Checkouts
  /customers:
    get:
      consumes:
      - application/json
      description: Get details of all customers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/usecase.CustomerFindAllOutputDTO'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all customers
      tags:
      - Customers
    post:
      consumes:
      - application/json
      description: Create Customer in DB
      parameters:
      - description: cria novo cliente
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/usecase.CustomerRegisterInputDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usecase.CustomerRegisterOutputDTO'
        "400":
          description: invalid data
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: customer already exists
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create Customer
      tags:
      - Customers
  /customers/{cpf}:
    get:
      consumes:
      - application/json
      description: Get details of a customer by cpf
      parameters:
      - description: Customer cpf
        in: path
        name: cpf
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usecase.CustomerFindByCpfOutputDTO'
        "404":
          description: Customer not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a customer
      tags:
      - Customers
    put:
      consumes:
      - application/json
      description: Update details of a customer by cpf
      parameters:
      - description: Customer cpf
        in: path
        name: cpf
        required: true
        type: string
      - description: Customer data
        in: body
        name: Customer
        required: true
        schema:
          $ref: '#/definitions/usecase.CustomerUpdateInputDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usecase.CustomerUpdateOutputDTO'
        "400":
          description: Invalid data
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Customer not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a customer
      tags:
      - Customers
  /kitchens/orders/cooking:
    post:
      consumes:
      - application/json
      description: Busca a ordem e o ticket da cozinha; move o status para o próximo
        da fase de preparo e faz o delivery se estiver finalizado.
      parameters:
      - description: indica a ordem a ser trabalhada pela cozinha
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/usecase.KitchenCookingInputDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/usecase.KitchenCookingOutputDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Busca a ordem e o ticket da cozinha; move o status para o próximo da
        fase de preparo e faz o delivery se estiver finalizado.
      tags:
      - Kitchens
...
  /kitchens/orders/flow:
    get:
      consumes:
      - application/json
      description: Retorna as ordens que estão na cozinha em alguma etapa do preparo
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/usecase.KitchenFindAllAOutputDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: retorna todas as ordens que estão na cozinha
      tags:
      - Kitchens
...
swagger: "2.0"